Clean up, clean up, everybody everywhere. Clean up, clean up, everybody do your share.
