+++ images/rpg/emptymap.png
