+++ pyproject.toml
@@ -12,6 +12,8 @@ flask = "^1.1.2"
 pymongo = "^3.11.3"
 dnspython = "^2.1.0"
 dadjokes = "^1.3.2"
+discord-pretty-help = "^1.2.2"
+pytz = "^2021.1"
 
 [tool.poetry.dev-dependencies]
